Fixed bug with add_task(), remove_task(), and display() functions.
The error could be replicated by first adding more than two tasks,
then removing the last task which would leave an improperly initialized
node in the "List" linked list. Then using the display_list() function
would cause the program to never find the "NULL" conditional for the
while loop in the function.

This issue was fixed by ackowledging that the remove_task() function
was not doing what was intended:

[*link_ref current]--[Target]--NULL

Instead of moving current to the correct location on the list:

[*link_ref]--[current]--NULL